Ball vs Boundaries
Determined that the code to prevent teleporting outside the bounds of the factory is somehow preventing interaction with the ball. Code for teleportation check is present in the LeftInput.cs file but is commented out for further inspection at a later time. At least this way demonstration can be had for UI/UX testing of the game functionality.